Move authenticate_subscription. Config -> Pusher
